New functions *.new() and ur()
   - added new function hitro.new().
   - added new functions dau.new() and dgt.new()
   - added new function ars.new()
   - added new function tdr.new()
   - added new function ur() (shorter alias for unuran.sample)
   - added man pages and tests for new functions

   - added new file 'R/universal.R' for functions that create
     unuran generator objects for particular methods
   - mark functions urtdr(), urhitro(), urdgt(), and urdau()
     as deprecated
   - Update man pages
   - Update test suite

+++ Runuran/R/unuran_discr.R
@@ -46,7 +46,7 @@ setMethod( "initialize", "unuran.discr",
                   ## Check entries
                   if(! (is.numeric(lb) && is.numeric(ub) && lb < ub) )
                           stop("invalid domain ('lb','ub')", call.=FALSE)
-                  if (!(is.double(pv) || is.null(pv) ))
+                  if (!(is.numeric(pv) || is.null(pv) ))
                           stop("invalid argument 'pv'", call.=FALSE)
                   if(! (is.function(pmf) || is.null(pmf)) )
                           stop("invalid argument 'pmf'", call.=FALSE)
@@ -54,7 +54,7 @@ setMethod( "initialize", "unuran.discr",
                   ## Store informations (if provided)
                   if (is.function(pmf)) .Object@pmf  <- pmf
                   ## (There is no need to store the PV)
-                  
+
                   ## We need an evironment for evaluating R expressions
                   .Object@env <- new.env()
